% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_df_to_list}
\alias{convert_df_to_list}
\title{Convert a data frame to a named list}
\usage{
convert_df_to_list(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A list, one element per row
}
\description{
Useful when converting to yaml.
}
\examples{
convert_df_to_list(dplyr::starwars)
}
