% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_list_to_df1}
\alias{convert_list_to_df1}
\title{Convert a list with single entries to a data frame}
\usage{
convert_list_to_df1(my_list)
}
\arguments{
\item{my_list}{A list with single entries}
}
\value{
A tibble with \code{key} and \code{value} columns
}
\description{
Convert a list with single entries to a data frame
}
\examples{
convert_list_to_df1(as.list(dplyr::starwars)[2])
}
