% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_list_to_df2}
\alias{convert_list_to_df2}
\title{Convert a list of lists to a data frame}
\usage{
convert_list_to_df2(my_list, as_character = TRUE, on_empty = NA)
}
\arguments{
\item{my_list}{A list of lists}

\item{as_character}{Logical; read all values as character}

\item{on_empty}{Value to return if \code{my_list} is NULL, NA or length zero}
}
\value{
A tibble, or \code{on_empty}
}
\description{
Requires that every list have the same named elements.
}
\examples{
convert_list_to_df2(convert_df_to_list(dplyr::starwars))
}
