% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{dataset_build}
\alias{dataset_build}
\title{Build dataset}
\usage{
dataset_build(
  filename_metadata,
  filename_data_raw,
  definitions,
  unit_conversion_functions,
  schema,
  resource_metadata,
  taxon_list,
  filter_missing_values = TRUE
)
}
\arguments{
\item{filename_metadata}{Metadata yaml file for a given study}

\item{filename_data_raw}{Raw \code{data.csv} file for any given study}

\item{definitions}{Definitions read in from the \code{traits.yml}}

\item{unit_conversion_functions}{\code{unit_conversion.csv} file read in from the config folder}

\item{schema}{Schema for responses.build}

\item{resource_metadata}{metadata for the compilation}

\item{taxon_list}{Taxon list}

\item{filter_missing_values}{Default filters missing values from the excluded data table;
change to false to see the rows with missing values}
}
\value{
List, AusTraits database object
}
\description{
Build specified dataset. This function completes three steps, which can be executed separately if desired:
\code{dataset_configure}, \code{dataset_process}, \code{dataset_update_taxonomy}
}
\examples{
\dontrun{
dataset_build(
  "data/Falster_2003/data.csv",
  "data/Falster_2003/metadata.yml",
  read_yaml("config/traits.yml"),
   get_unit_conversions("config/unit_conversions.csv"),
   get_schema(),
   get_schema("config/metadata.yml", "metadata"),
   read_csv_char("config/taxon_list.csv")
)
}
}
