% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{dataset_test_worker}
\alias{dataset_test_worker}
\title{Test whether specified \code{dataset_id} has the correct setup}
\usage{
dataset_test_worker(
  test_dataset_ids,
  path_config = "config",
  path_data = "data",
  schema = get_schema(),
  definitions = get_definitions(dir = path_config)
)
}
\arguments{
\item{test_dataset_ids}{Vector of \code{dataset_id} for sources to be tested}

\item{path_config}{Path to folder containing configuration files}

\item{path_data}{Path to folder containing data files}

\item{schema}{Data schema}

\item{definitions}{Trait defininitons}
}
\description{
Run tests to ensure that specified \code{dataset_id} has the correct setup.
}
