% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.build-package.R
\docType{package}
\name{responses.build-package}
\alias{responses.build}
\alias{responses.build-package}
\title{responses.build: A Workflow for Harmonising Physiological Response-Curve Data into a Documented Standard Structure}
\description{
The `responses.build` package provides a workflow to harmonise physiological response-curve data - leaf gas exchange, hydraulic vulnerability, and fluorescence response curves - from diverse sources into a documented, relational structure. It is a fork of `traits.build` (Wenk et al, 2024, \doi{10.1016/j.ecoinf.2024.102773}, \url{https://github.com/traitecoevo/traits.build}), which harmonises single-valued trait data; this package extends the same data model to measurements that form curves, where a response is measured across a driver gradient and the individual points belong to an identifiable curve. It was built to support AusFizz (\url{https://github.com/traitecoevo/AusFizz}). For the shared data model and workflow concepts see \url{https://traitecoevo.github.io/traits.build-book/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://traitecoevo.github.io/responses.build/}
  \item Report bugs at \url{https://github.com/traitecoevo/responses.build/issues}
}

}
\author{
\strong{Maintainer}: Daniel Falster \email{daniel.falster@unsw.edu.au} (\href{https://orcid.org/0000-0002-9814-092X}{ORCID})

Authors:
\itemize{
  \item Daniel Falster \email{daniel.falster@unsw.edu.au} (\href{https://orcid.org/0000-0002-9814-092X}{ORCID})
  \item Elizabeth Wenk (\href{https://orcid.org/0000-0001-5640-5910}{ORCID}) [curator]
  \item David Coleman \email{david.coleman@mq.edu.au} [curator]
}

Other contributors:
\itemize{
  \item Sophie Yang (\href{https://orcid.org/0000-0001-7328-345X}{ORCID}) [contributor]
  \item Fonti Kar (\href{https://orcid.org/0000-0002-2760-3974}{ORCID}) [contributor]
  \item ARDC [funder]
  \item ARC [funder]
}

}
\keyword{internal}
