% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util_df_convert_character}
\alias{util_df_convert_character}
\title{Convert all columns in data frame to character}
\usage{
util_df_convert_character(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe
}
\description{
Convert all columns in data frame to character
}
\examples{
lapply(responses.build:::util_df_convert_character(dplyr::starwars), class)
}
